requirements.txt
openai-whisper
streamlit
pydub